Centra la seccion de habilidades

+++ views/users/progreso.php
@@ -1,10 +1,11 @@
-<?php 
+<?php
+
 /** @var mysqli_result $materias */
-?>  
+?>
 <section id="progreso">
     <h1>Mi progreso en la Tecnictura Universitaria en Programación</h1>
 
-    <div class="center tableContainer">
+    <div class="center centeredContainer">
         <table>
             <thead>
                 <tr>
@@ -17,7 +18,7 @@
                 </tr>
             </thead>
             <tbody>
-                <?php while($materia = $materias->fetch_assoc()): ?>
+                <?php while ($materia = $materias->fetch_assoc()): ?>
                     <tr>
                         <td><?= htmlspecialchars($materia["materia_id"]) ?></td>
                         <td><?= htmlspecialchars($materia["nombre"]) ?></td>
@@ -25,11 +26,11 @@
                         <td><?= htmlspecialchars($materia["cuatrimestre"]) ?></td>
                         <td class="estado-<?= htmlspecialchars($materia["estado"]) ?>"><?= htmlspecialchars($materia["estado"]) ?></td>
                         <td>
-                            <a class="actionBtn edit"  href="index.php?action=edit&materia_id=<?=$materia["materia_id"] ?>">Editar</a>
-                            <a class="actionBtn delete" href="index.php?action=delete&materia_id=<?=$materia["materia_id"] ?>">Eliminar</a>
+                            <a class="actionBtn edit" href="index.php?action=edit&materia_id=<?= $materia["materia_id"] ?>">Editar</a>
+                            <a class="actionBtn delete" href="index.php?action=delete&materia_id=<?= $materia["materia_id"] ?>" onclick="return confirm('¿Eliminar este usuario?')">Eliminar</a>
                         </td>
                     </tr>
-                <?php endwhile;?>
+                <?php endwhile; ?>
             </tbody>
         </table>
         <p><a href="index.php?action=create" class="actionBtn add">Agregar materia</a></p>
